refactor(database): split monolithic sqlx adapter into modular submodules
- Move database adapter from single file to mod.rs with organized submodules
- Extract account operations into dedicated account_ops.rs module
- Extract API key operations into dedicated api_key_ops.rs module
- Extract invitation operations into dedicated invitation_ops.rs module
- Extract member operations into dedicated member_ops.rs module
- Extract organization operations into dedicated organization_ops.rs module
- Extract passkey operations into dedicated passkey_ops.rs module
- Extract session operations into dedicated session_ops.rs module
- Extract two-factor operations into dedicated two_factor_ops.rs module
- Extract user operations into dedicated user_ops.rs module
- Extract verification operations into dedicated verification_ops.rs module
- Add SQLite support alongside existing PostgreSQL adapter
- Add sqlx-sqlite feature flag to workspace and core crates
- Include sqlite in sqlx dependency features for broader database compatibility
- Improves code organization and maintainability for database adapter implementation
