+++ Cargo.toml
@@ -35,6 +35,7 @@ async-trait = "0.1"
 sqlx = { version = "0.9.0", features = [
     "runtime-tokio",
     "postgres",
+    "sqlite",
     "chrono",
     "uuid",
     "json",
@@ -97,6 +98,7 @@ rustls = ["better-auth-api/rustls"]
 axum = ["dep:axum", "dep:tower", "dep:url", "better-auth-core/axum"]
 derive = ["better-auth-core/derive"]
 sqlx-postgres = ["dep:sqlx", "better-auth-core/sqlx-postgres"]
+sqlx-sqlite = ["dep:sqlx", "better-auth-core/sqlx-sqlite"]
 redis-cache = ["dep:redis"]
 
 [dependencies]
